закглушка для создания заказов из корзины БЕЗ ОПЛАТЫ изменение профиля заказчика обновление данных заказчика после оформления заказа

+++ src/main/java/ru/platform/user/dto/response/CustomerProfileRsDto.java
@@ -37,7 +37,7 @@ public class CustomerProfileRsDto {
     private String imageUrl;
 
     @Schema(description = "Процент скидки пользователя", example = "1")
-    private Integer discountPercentage;
+    private BigDecimal discountPercentage;
 
     @Schema(description = "Статус пользователя", example = "EXPLORER", enumAsRef = true)
     private CustomerStatus status;
